+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## 2.0.0
+### Added
+- Option to merge file contents line by line, this is useful for .gitignore files
+- Option to force overwrite files during installation
+
 ### Removed
 - Support for PHP < 8.1